requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
drf-yasg>=1.21
psycopg2-binary>=2.9      # Only required for PostgreSQL, can be omitted for SQLite
coverage>=7.2